backend/requirements.txt
astroid==1.6.4
certifi==2018.4.16
chardet==3.0.4
click==6.7
Flask==1.0.2
Flask-Cors==3.0.4
future==0.16.0
idna==2.6
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
pycodestyle==2.3.1
pyflakes==1.6.0
PyJWT==1.6.4
python-dotenv==0.8.2
python-telegram-bot==10.1.0
requests==2.18.4
six==1.11.0
urllib3==1.22
Werkzeug==0.14.1
wrapt==1.10.11
gunicorn